+++ lib/components/Loader.vue
@@ -1,0 +1,30 @@
+<template>
+  <div class="dgv-loader-container">
+    <div class="dgv-loader">
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+    </div>
+  </div>
+</template>
+
+<script lang="ts">
+import { defineComponent } from 'vue'
+
+/**
+ * @group Components
+ * @description The loader that is displayed when the data grid is loading page data.
+ */
+export default defineComponent({
+  name: 'Loader',
+})
+</script>